# 7th/cowork/swagger.json
{
  "swagger": "2.0",
  "schemes": [
    "http"
  ],
  "host": "220.230.114.8:3987",
  "basePath": "/",
  "info":
  {
    "description": "UNI SOPT 7차 개발 합동세미나 API Reference 입니다. <!-- ReDoc-Inject: <security-definitions> -->\n",
    "version": "1.0.0",
    "title": "UNI SOPT _ SERVER PART",
    "x-logo":
    {
      "url": "https://s3.ap-northeast-2.amazonaws.com/sopt21th/sopt21th_logo.png"
    }
  },
...
  "securityDefinitions":
  {
    "petstore_auth":
    {
      "description": "Get access to data while protecting your account credentials.\nOAuth2 is also a safer and more secure way to give you access.\n",
      "type": "oauth2",
      "authorizationUrl": "http://petstore.swagger.io/api/oauth/dialog",
      "flow": "implicit",
      "scopes":
      {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    },
    "api_key":
    {
      "description": "For this sample, you can use the api key `special-key` to test the authorization filters.\n",
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "paths":
...
    "/signup":
    {
      "post":
      {
        "tags": [
          "sign"
        ],
        "summary": "회원가입",
        "description": "사용자 회원가입을 합니다. Request : multipart/form-data",
        "operationId": "sign up",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
        {
          "in": "body",
          "name": "body",
          "description": "유저 회원가입을 하기 위한 email, pwd, nickname, profileImage 전달",
          "required": true,
          "schema":
          {
            "type": "object",
            "properties":
            {
              "email":
              {
                "description": "유저 email",
                "type": "string",
                "minLength": 1,
                "required": true,
                "example": "test@sopt.org"
              },
              "pwd":
              {
                "description": "비밀번호",
                "type": "string",
                "minLength": 1,
                "required": true,
                "example": "qwer1234"
              },
              "nickname":
              {
                "description": "유저 닉네임",
                "type": "string",
                "minLength": 1,
                "required": true,
                "example": "테스트"
              },
              "image":
              {
                "description": "유저의 프로필로 등록할 이미지파일",
                "required": true,
                "type": "image"
              }
            }
          }
        }],
        "responses":
        {
          "201":
          {
            "description": "이용자 회원가입 성공",
            "schema":
            {
              "type": "object",
              "properties":
              {
                "status":
                {
                  "type": "string",
                  "description": "성공 알림",
                  "example": "success"
                },
                "msg":
                {
                  "type": "string",
                  "description": "회원가입 성공 알림 메세지",
                  "example": "successful signup"
                }
              }
            }
          },
          "401":
          {
            "description": "email overlap",
            "schema":
            {
              "$ref": "#/definitions/failMessage"
            }
          },
          "500":
          {
            "description": "get connection error // get user data error // hash password error // insert user data error",
            "schema":
            {
              "$ref": "#/definitions/failMessage"
            }
          }
        }
      }
    },
...
  },
  "definitions":
...
    "failMessage":
    {
      "type": "object",
      "properties":
      {
        "status":
        {
          "type": "string",
          "example": "fail"
        },
        "message":
        {
          "type": "string",
          "example": "fail reason"
        }
      }
    }
  }
}